requirements.txt
anyio==4.2.0
asgiref==3.7.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
contextlib2==21.6.0
crispy-bootstrap5==2023.10
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.2
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.1
djangorestframework==3.14.0
duplicity==0.8.21
fasteners==0.14.1
future==0.18.2
graphql-core==3.2.3
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libcst==1.1.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-multipart==0.0.9
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
rename==22.12.0
reportlab==3.6.8
requests==2.25.1
rich==13.7.0
schema==0.7.5
SecretStorage==3.3.1
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.37.1
strawberry-graphql==0.219.2
strawberry-graphql-django==0.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ufw==0.36.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.27.1
wadllib==1.3.6
wheel==0.42.0
xdg==5
zipp==1.0.0
